ChatAgent/requirements.txt
mem0ai
openai
python-dotenv
chromadb